refactor: 투표 메서드 인자 수정
