+++ app/src/main/kotlin/com/wespot/vote/VoteController.kt
@@ -1,13 +1,11 @@
 package com.wespot.vote
 
-import com.wespot.vote.dto.VoteItems
+import com.wespot.vote.dto.request.VoteRequest
+import com.wespot.vote.dto.response.VoteItems
 import com.wespot.vote.port.`in`.VoteUseCase
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 
 @RestController
 @RequestMapping("/votes")
@@ -23,8 +21,8 @@ class VoteController(
     }
 
     @PostMapping
-    fun createVote(userId: Long): ResponseEntity<Long> {
-        val savedId: Long = voteUseCase.saveVote(userId)
+    fun createVote(userId: Long, @RequestBody requests: List<VoteRequest>): ResponseEntity<Long> {
+        val savedId: Long = voteUseCase.saveVote(userId, requests)
         return ResponseEntity.status(HttpStatus.CREATED)
             .body(savedId)
     }
