+++ core/src/main/kotlin/com/wespot/vote/dto/request/VoteRequest.kt
@@ -1,0 +1,7 @@
+package com.wespot.vote.dto.request
+
+class VoteRequest(
+    val userId: Long,
+    val voteOptionId: Long
+) {
+}
