+++ core/src/main/kotlin/com/wespot/vote/port/in/VoteUseCase.kt
@@ -1,6 +1,7 @@
 package com.wespot.vote.port.`in`
 
-import com.wespot.vote.dto.VoteItems
+import com.wespot.vote.dto.request.VoteRequest
+import com.wespot.vote.dto.response.VoteItems
 import org.springframework.stereotype.Service
 
 @Service
@@ -8,6 +9,6 @@ interface VoteUseCase {
 
     fun getVoteOptions(userId: Long): VoteItems
 
-    fun saveVote(userId: Long): Long
+    fun saveVote(userId: Long, requests: List<VoteRequest>): Long
 
 }
